Moved implicit conversions to separated Command.
