+++ lib/naught/null_class_builder/commands.rb
@@ -1,0 +1,2 @@
+require 'naught/null_class_builder/commands/define_explicit_conversions'
+require 'naught/null_class_builder/commands/define_implicit_conversions'
