+++ lib/naught/null_class_builder/commands/define_implicit_conversions.rb
@@ -1,0 +1,20 @@
+require 'naught/null_class_builder/commands/base'
+
+module Naught
+  class NullClassBuilder
+    module Commands
+      class DefineImplicitConversions < ::Naught::NullClassBuilder::Commands::Base
+
+        def call
+          defer do |subject|
+            subject.module_eval do
+              def to_ary; []; end
+              def to_str; ''; end
+            end
+          end
+        end
+
+      end
+    end
+  end
+end
